Heading and style cleanup + started work on cards
The headings have been changed from "About Me", "My Games", and "My Web Experiments" to "About Me", "About This Website", and "My Stuff". This change has been reflected in the navigation bar as well.

Started work on link cards and filtering. Scripts have been split so that the age calculation is done in the html in a script tag, the hero section and filtering will have their own scripts.